start, lab #1
This first project uses the STM32L-Discovery board
[http://www.st.com/internet/evalboard/product/250990.jsp]
and displays a counter using the LCD while also blinking
the LED on pin PB6.  The main body is written in C but
it includes sections written in assembly.

The "empty_project" contains a starting project for
these which are developed with the IAR embedded workbench
[www.iar.com].  Each derived project does not re-include
all these files so they must copied in to each directory in
order for them to compile properly.