reliable SPI communications complete!
Fixed CPLD 'reset' has no load problem.

For the CPLD, added 'main.xcf' for
the Diamond Lattice programmer.

The ARM has been coded to display the data
it receives on the LCD and also send
it back to the CPLD which displays
it on the on board LEDs.
With this working properly the on board LEDs
match the external LEDs exactly.