A blizzard of synthesis errors
Fixed synthesis problems with switch_ctl.v

Both spi_ctl.v and mem_ctl.v have lots of synthesis
errors and need to be re-written.
When level triggered @(var) it wants to include every
possible variable.
When edge triggered it can't be triggered on both edges.
And on and on.