memory writes working, but not perfect
Got the memory writes working by adding a delay
to the end of the SPI_once function.
This is a bit of hack and more investigation is
needed to determine why this works and if this
is the best solution.

Also the values written are inverted compared to
those read (e.g. writing 0x00 reads 0xFF).