fixed writes involving led_ctl module
Changed to latch data leaving write
cycle as opposed to entering
(posedge of write_n instead of negedge).

Removed the default case for write_data_bus in spi_ctl.
This should cause it to become a "latch" so that it
will retain its value.  The Synplify synthesis tool
will warn about this but it is acceptable.